requirements.txt
# D-Wave Ocean SDK Requirements
dwave-ocean-sdk>=6.0.0
dimod>=0.12.0
dwave-system>=1.20.0
dwave-samplers>=1.0.0
dwavebinarycsp>=0.2.0
networkx>=2.5
numpy>=1.20.0
matplotlib>=3.3.0
jupyter>=1.0.0